+++ src/App.test.tsx
@@ -1,4 +1,4 @@
-// import { render, screen } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import React from 'react'
 import ReactDOM from 'react-dom'
 import SamuraiJSApp from "./App";
@@ -8,6 +8,6 @@ test('renders learn react link', () => {
   ReactDOM.render(<SamuraiJSApp />,div);
   ReactDOM.unmountComponentAtNode(div);
 
-  // const linkElement = screen.getByText(/learn react/i);
-  // expect(linkElement).toBeInTheDocument();
+  const linkElement = screen.getByText(/learn react/i);
+  expect(linkElement).toBeInTheDocument();
 });
