+++ src/components/Users/UsersContainer.tsx
@@ -1,99 +1,22 @@
-import React from "react";
-import {connect} from "react-redux";
-import Users from "./Users";
-import Preloader from "../commons/Preloader/Preloader";
-import {
-    follow, unfollow, requestUsers, FilterType
-} from '../redux/users-reducer'
-import {compose} from "redux";
-import {
-    getCurrentPage,
-    getFollowingInProgress,
-    getIsFetching,
-    getPageSize,
-    getTotalUsersCount,
-    getUsers, getUsersFilter
-} from '../redux/users-selector'
-import {UserType} from "../../types/types";
-import {AppStateType} from "../redux/redux-store";
-import {RouteComponentProps} from "react-router-dom";
-import {withAuthRedirect} from "../../hoc/withAuthRedirect";
+import React from 'react'
+import {useSelector} from 'react-redux'
+import {Users} from './Users'
+import Preloader from '../commons/Preloader/Preloader'
+import {getIsFetching, getTotalUsersCount} from '../redux/users-selector'
 
-type MapStatePropsType = {
-    currentPage: number
-    pageSize: number
-    isFetching: boolean
-    totalUsersCount: number
-    users: Array<UserType>
-    followingInProgress: Array<number>
-    filter: FilterType
-}
-type MapDispatchPropsType = {
-    getUsers: (currentPage: number, pageSize: number, filter: FilterType) => void
-    follow: (userId: number) => void
-    unfollow: (userId: number) => void
-}
-type OwnPropsType = {
+
+type PropsType = {
     pageTitle: string
 }
-type PropsType = MapStatePropsType & MapDispatchPropsType & OwnPropsType
 
-class UsersContainer extends React.Component<PropsType, any> {
+export const UsersPage: React.FC<PropsType> = (props) => {
 
-    componentDidMount() {
-        const {currentPage, pageSize, getUsers, filter} = this.props
-        getUsers(currentPage, pageSize, filter)
-    }
+    const isFetching = useSelector(getIsFetching)
+    const totalUsersCount = useSelector(getTotalUsersCount)
 
-    onPageChanged = (pageNumber: number) => {
-        const {pageSize, getUsers, filter} = this.props
-        getUsers(pageNumber, pageSize, filter)
-    }
-
-    onFilterChanged = (filter: FilterType) => {
-        const {pageSize} = this.props
-        this.props.getUsers(1, pageSize, filter)
-    }
-
-    render() {
-        return <>
-            <h2>{this.props.pageTitle}({this.props.totalUsersCount})</h2>
-            {this.props.isFetching ? <Preloader/> : null}
-            <Users
-                users={this.props.users}
-                totalUsersCount={this.props.totalUsersCount}
-                pageSize={this.props.pageSize}
-                currentPage={this.props.currentPage}
-                onPageChanged={this.onPageChanged}
-                onFilterChanged={this.onFilterChanged}
-                unfollow={this.props.unfollow}
-                follow={this.props.follow}
-                followingInProgress={this.props.followingInProgress}
-            />
-        </>
-    }
+    return <>
+        <h2>{props.pageTitle}({totalUsersCount})</h2>
+        {isFetching ? <Preloader/> : null}
+        <Users/>
+    </>
 }
-
-const mapStateToProps = (state: AppStateType): MapStatePropsType => {
-    return {
-        users: getUsers(state),
-        pageSize: getPageSize(state),
-        totalUsersCount: getTotalUsersCount(state),
-        currentPage: getCurrentPage(state),
-        isFetching: getIsFetching(state),
-        followingInProgress: getFollowingInProgress(state),
-        filter: getUsersFilter(state)
-    }
-}
-
-export default compose(
-  //withAuthRedirect,
-    //TStateProps = {}, TDispatchProps = {}, TOwnProps = {}, State = DefaultState
-    connect
-    <MapStatePropsType, MapDispatchPropsType, OwnPropsType, AppStateType>
-    (mapStateToProps, {
-        follow,
-        unfollow,
-        getUsers: requestUsers
-    })
-)(UsersContainer)
